+++ ContinuousStepper_Timer3.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include <TimerThree.h> // https://github.com/PaulStoffregen/TimerThree
+
+#include "ContinuousStepper_Timer.h"
+
+typedef ContinuousStepper_Timer<TimerThree> ContinuousStepper_Timer3;