+++ CHANGELOG.md
@@ -5,6 +5,7 @@ HEAD
 ----
 
 * Fix error when the `TimerOne` library is not available
+* Add class `ContinuousStepper_Timer3` to support the `TimerThree` library
 
 1.1.0
 -----